Changed how IP Address of Client is entered
Instead of changing the IP Address of the client within the code constantly, the program is now set up so that a JOptionPane asks the user to input the Client's IP Address. This way the program can run on a desktop and the user never has to have access to the code to change the IP Address.